Allow to override build date with SOURCE_DATE_EPOCH
to make builds reproducible.
See https://reproducible-builds.org/ for why this is good
and https://reproducible-builds.org/specs/source-date-epoch/
for the definition of this variable.

Also use UTC/gmtime to be independent of timezone.

This patch was done while working on reproducible builds for openSUSE, sponsored by the NLnet NGI0 fund.